requirements.txt
# === Core ===
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.3.2
joblib==1.3.2
cmake==3.30.5
lightgbm==4.5.0
scipy==1.11.4

# === NLP ===
nltk==3.8.1
gensim==4.3.2
sentence-transformers==2.2.2
stopwordsiso==0.6.1
swifter==1.3.5
huggingface_hub==0.10.1
# === Oversampling ===
imbalanced-learn==0.11.0

# === Similarity ===
faiss-cpu==1.7.4

# === API ===
fastapi==0.111.0
uvicorn==0.30.1

# === Visualization (Optional) ===
matplotlib==3.7.3
seaborn==0.13.0
plotly==5.19.0
